Refactored IIN agent base server code
